07/2563

+++ pensso/src/com/isecinc/pens/bean/Contact.java
@@ -1,6 +1,7 @@
 package com.isecinc.pens.bean;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 
 import util.ConvertNullUtil;
@@ -33,7 +34,7 @@ public class Contact extends I_PO implements Serializable {
 	 */
 	public Contact(ResultSet rst) throws Exception {
 		setId(rst.getInt("CONTACT_ID"));
-		setCustomerId(rst.getInt("CUSTOMER_ID"));
+		setCustomerId(rst.getLong("CUSTOMER_ID"));
 		setContactTo(rst.getString("CONTACT_TO").trim());
 		setRelation(ConvertNullUtil.convertToString(rst.getString("RELATION")).trim());
 		setPhone(rst.getString("PHONE").trim());
@@ -85,7 +86,7 @@ public class Contact extends I_PO implements Serializable {
 	private String isActive;
 
 	/** CUSTOMER ID */
-	private int customerId;
+	private long customerId;
 
 	/** PHONE */
 	private String phone2;
@@ -150,11 +151,11 @@ public class Contact extends I_PO implements Serializable {
 		this.isActive = isActive;
 	}
 
-	public int getCustomerId() {
+	public long getCustomerId() {
 		return customerId;
 	}
 
-	public void setCustomerId(int customerId) {
+	public void setCustomerId(long customerId) {
 		this.customerId = customerId;
 	}
